Moved GameArea file to core folder
